Moved loading of plugin translations from Plugin Manager to Translator
